@@ -1,12 +1,13 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
+
     defaultConfig {
         minSdkVersion 23
-        targetSdkVersion 28
-        versionCode 2
-        versionName "1.11"
+        targetSdkVersion 29
+        versionCode 3
+        versionName "2.00"
     }
     buildTypes {
         release {
@@ -18,5 +19,6 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.2.0-alpha05'
 }
