@@ -1,18 +1,12 @@
 package com.gocalsd.switchbarexample;
 
-import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.widget.CompoundButton;
-import android.widget.Switch;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.gocalsd.switchbar.SwitchBar;
+import com.gocalsd.switchbar.ToggleSwitch;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -23,48 +17,34 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Toolbar toolbar = findViewById(R.id.toolbar);
+
+        final Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        posMessage = "Switchbar is Enabled!";
-        negMessage = "Switchbar Disabled...";
+        switchBar = new SwitchBar(this);
+
+        posMessage = getString(R.string.switch_on_string);
+        negMessage = getString(R.string.switch_off_string);
 
         switchBar = findViewById(R.id.switchBar);
-        //Show the Switchbar! It's hidden by default!
-        switchBar.show();
-
-        //Set the on and off Messages
-        switchBar.setOnMessage(posMessage);
-        switchBar.setOffMessage(negMessage);
 
         //Set the on and off Background Colors
         //If you want to theme the Switch, override the style
         switchBar.setSwitchbarOnBackground(R.color.colorPrimary);
-        switchBar.setSwitchbarOffBackground(R.color.colorPrimaryDark);
+        switchBar.setSwitchbarOffBackground(R.color.colorAccent);
 
         //Set the defaults
+        switchBar.setOnMessage(posMessage);
+        switchBar.setOffMessage(negMessage);
+
         switchBar.setChecked(true);
         switchBar.setTextViewLabel(true);
 
         //Set the addOnSwitchChangeListener
         switchBar.addOnSwitchChangeListener(new SwitchBar.OnSwitchChangeListener() {
             @Override
-            public void onSwitchChanged(Switch switchView, boolean isChecked) {
-                if(isChecked){
-                    Toast.makeText(MainActivity.this, "It's on!", Toast.LENGTH_SHORT).show();
-                }else{
-                    Toast.makeText(MainActivity.this, "It's off!", Toast.LENGTH_SHORT).show();
+            public void onSwitchChanged(ToggleSwitch switchView, boolean isChecked) {
 
-                }
-            }
-        });
-
-        FloatingActionButton fab = findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
             }
         });
     }
